requirements.txt
chromadb
langchain-chroma
langchain-openai
langchain-text-splitters
langchain-community
streamlit
transformers
PyMuPDF
pillow
open_clip_torch
torch